initial stab at a binding.gyp file
Kind of brain-dead gyp file; arranges to run gsl's configure and
make, which is total overkill for the handful of functions in this
package.  But a learning exercise.

Currently `node-gyp configure` and `node-gyp build` will end up
running gsl's `make`, but the link step with the node shims is
failing.  Good enough for initial commit!